+++ CHANGES.rst
@@ -18,6 +18,9 @@ New Features
 
 - ``astropy.io.fits``
 
+  - Fixed an bug where updates to string columns in FITS tables were not saved
+    on Python 3. [#4452]
+
 - ``astropy.io.misc``
 
 - ``astropy.io.registry``
@@ -808,6 +811,9 @@ Bug Fixes
 - ``astropy.io.ascii``
 
 - ``astropy.io.fits``
+
+  - Fixed an bug where updates to string columns in FITS tables were not saved
+    on Python 3. [#4452]
 
 - ``astropy.io.misc``